Refactor ToGraphviz and ToMermaid classes to reduce code duplication
