+++ src/test/scala/cego/GraphvizTest.scala
@@ -1,18 +1,19 @@
 package cego
 
 import cego.rbtX.{ToRbt, FromRbt}
-import cego.graphvizX.*
 import cego.scalaX.*
+
 class GraphvizTest extends RBTSpec {
+
   it should "small" in {
     val imp = FromRbt.load(s"$TST_RES/small.rbt")
     val snap = imp.snap
     ToRbt(snap).toXML
     val xml = ToRbt(snap).toXmlString
-    xml.writeTo( s"$TARGET/Generated.rbt" )
+    xml.writeTo(s"$TARGET/Graphviz.rbt")
     Root.restore(snap)
     val code = ToScala(snap).toString
-    code.writeTo(s"$TARGET/Generated.txt")
-    imp.toAdoc.writeTo(s"$TARGET/Generated.adoc")
+    code.writeTo(s"$TARGET/Graphviz.txt")
+    imp.toAdoc.writeTo(s"$TARGET/Graphviz.adoc")
   }
 }
